refactor: consolidate common test fixtures
